Add files via upload
This is a fully working implementation of my fan controller. It starts automatically at boot or re-load and runs forever.
There are no GUI attachments or elements, although I do use the CPU temperature card to keep track of the temperature over time.
This uses a very simple method to drive the fan. It provides sufficient cooling so the RPi will not throttle the speed.